requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.6
chardet==4.0.0
discord.ext.context==0.1.7
discord.py==1.7.3
idna==3.2
multidict==5.1.0
mutagen==1.45.1
pycparser==2.20
pycryptodomex==3.11.0
PyNaCl==1.4.0
six==1.16.0
typing-extensions==3.10.0.2
websockets==10.0
yarl==1.6.3
youtube-dl==2021.6.6
yt-dlp==2021.10.10
